style(header): header in phone version fully redesigned
