+++ src/Components/header/Header.jsx
@@ -1,139 +1,14 @@
-"use client";
-
-export const metadata = {
-  title: "Crazy Ecommerce",
-  description: "Generated by create next app",
-};
-import React, { useState } from "react";
-import Cyber from "../../../assets/cyberIcon";
-import {
-  SearchIcon,
-  Heart,
-  ShoppingCart,
-  CircleUserRound,
-  Menu,
-} from "lucide-react";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { useTranslations } from "next-intl";
-import {
-  SignIn,
-  SignInButton,
-  SignUpButton,
-  SignedIn,
-  SignedOut,
-  UserButton,
-  useUser
-} from "@clerk/nextjs";
-
-// import LangSwitcher from "@/i18n/Switcher"
-
-const routes = [
-  { name: "Home", path: "/" },
-  { name: "Catalog", path: "/catalog" },
-  { name: "About", path: "/about" },
-  { name: "Contact us", path: "/contact" },
-  { name: "Wishlist", path: "/wishlist" , onlyPhone: true },
-  { name: "Cart", path: "/cart" , onlyPhone: true },
-];
+import React from 'react'
+import DesktopHeader from './desktop/DesktopHeader'
+import MobileHeader from './mobile/MobileHeader'
 
 const Header = () => {
-  const t = useTranslations();
-  const [open, setOpen] = useState(false);
-  const pathname = usePathname();
-
   return (
-    <div className="fixed bottom-0 left-0 h-fit w-full z-[9999] bg-white shadow-sm md:top-0">
-      <header className="flex items-center justify-between gap-14 p-4 md:justify-center h-22 text-black relative">
-        {/* Company logo */}
-        <div className="hidden">{<Cyber />}</div>
+    <>
+        <DesktopHeader/>
+        <MobileHeader/>
+    </>
+  )
+}
 
-        {/* Search input */}
-        <div className="hidden p-4 gap-8 rounded-2xl bg-search-bg w-full md:flex md:max-w-93">
-          <SearchIcon />
-          <input
-            type="text"
-            placeholder={t("header.search_placeholder")}
-            className="w-full bg-transparent focus:outline-none"
-          />
-        </div>
-
-        {/* Desktop nav */}
-        <ul className="hidden md:flex gap-13">
-          {routes.map((route) => ( !route.onlyPhone &&
-            <li
-              className={pathname === route.path ? "font-bold" : ""}
-              key={route.name}
-            >
-              <Link href={route.path}>{t(`header.nav.${route.name}`)}</Link>
-            </li>
-          ))}
-        </ul>
-
-        {/* Icons */}
-        <div className="hidden md:flex gap-6">
-          {/* <LangSwitcher /> */}
-          <Link href="/Wishlist">
-            <Heart />
-          </Link>
-          <Link href="/cart">
-            <ShoppingCart />
-          </Link>
-          <SignedIn>
-            <UserButton />
-          </SignedIn>
-          <SignedOut>
-            <SignUpButton mode="modal" children={<CircleUserRound />} />
-          </SignedOut>
-        </div>
-
-        {/* Mobile menu button */}
-        {/* <div>
-          <Menu
-            onClick={() => setOpen(!open)}
-            className="md:hidden cursor-pointer"
-          />
-        </div> */}
-      </header>
-
-      {/* Mobile menu */}
-       
-        {/* <div className="absolute top-full left-0 w-full bg-white z-[9998] border-t transition-all md:hidden">
-          <div className="flex center">
-            <SignedIn>
-              <UserButton showName />
-            </SignedIn>
-          </div>
-          <div className="flex">
-            <SignedOut>
-              <div className="btn border w-full text-black">
-                <SignInButton mode="modal" />
-              </div>
-              <button className="btn bg-black w-full border text-white">
-                <SignUpButton mode="modal" />
-              </button>
-            </SignedOut>
-          </div>
-          <div className="py-4 flex flex-col gap-1">
-            {routes.map((route) => (
-              <li
-                className={pathname === route.path ? "font-bold" : ""}
-                key={route.name}
-              >
-                <Link
-                  onClick={() => setOpen(false)}
-                  className="flex justify-center text-xl py-3"
-                  href={route.path}
-                >
-                  {route.name}
-                </Link>
-              </li>
-            ))}
-          </div>
-        </div> */}
-    
-    </div>
-  );
-};
-
-export default Header;
+export default Header
