+++ src/Components/header/desktop/DesktopHeader.jsx
@@ -1,0 +1,94 @@
+"use client";
+
+export const metadata = {
+  title: "Crazy Ecommerce",
+  description: "Generated by create next app",
+};
+import React, { useState } from "react";
+import Cyber from "../../../../assets/cyberIcon";
+import {
+  SearchIcon,
+  Heart,
+  ShoppingCart,
+  CircleUserRound,
+  Menu,
+} from "lucide-react";
+import Link from "next/link";
+import { usePathname } from "next/navigation";
+import { useTranslations } from "next-intl";
+import {
+  SignIn,
+  SignInButton,
+  SignUpButton,
+  SignedIn,
+  SignedOut,
+  UserButton,
+  useUser
+} from "@clerk/nextjs";
+
+// import LangSwitcher from "@/i18n/Switcher"
+
+const routes = [
+  { name: "Home", path: "/" },
+  { name: "Catalog", path: "/catalog" },
+  { name: "About", path: "/about" },
+  { name: "Contact us", path: "/contact" },
+  { name: "Wishlist", path: "/wishlist" , onlyPhone: true },
+  { name: "Cart", path: "/cart" , onlyPhone: true },
+];
+
+const DesktopHeader = () => {
+  const t = useTranslations();
+  const [open, setOpen] = useState(false);
+  const pathname = usePathname();
+
+  return (
+    <div className="hidden fixed md:block md:left-0 h-fit w-full z-[9999] bg-white shadow-sm md:top-0">
+      <header className="flex items-center justify-between gap-14 p-4 md:justify-center h-22 text-black relative">
+        {/* Company logo */}
+        <div className="hidden md:block">{<Cyber />}</div>
+
+        {/* Search input */}
+        <div className="hidden p-4 gap-8 rounded-2xl bg-search-bg w-full md:flex md:max-w-93">
+          <SearchIcon />
+          <input
+            type="text"
+            placeholder={t("header.search_placeholder")}
+            className="w-full bg-transparent focus:outline-none"
+          />
+        </div>
+
+        {/* Desktop nav */}
+        <ul className="hidden md:flex gap-13">
+          {routes.map((route) => ( !route.onlyPhone &&
+            <li
+              className={pathname === route.path ? "font-bold" : ""}
+              key={route.name}
+            >
+              <Link href={route.path}>{t(`header.nav.${route.name}`)}</Link>
+            </li>
+          ))}
+        </ul>
+
+        {/* Icons */}
+        <div className="hidden md:flex gap-6">
+          {/* <LangSwitcher /> */}
+          <Link href="/Wishlist">
+            <Heart />
+          </Link>
+          <Link href="/cart">
+            <ShoppingCart />
+          </Link>
+          <SignedIn>
+            <UserButton />
+          </SignedIn>
+          <SignedOut>
+            <SignUpButton mode="modal" children={<CircleUserRound />} />
+          </SignedOut>
+        </div>
+      </header>
+    </div>
+  );
+};
+
+export default DesktopHeader;
