+++ src/Components/ui/product/addtocart/components/counter.jsx
@@ -3,8 +3,8 @@ import { ShoppingCart, PlusSquare, MinusSquare, Trash2 } from "lucide-react";
 
 const Counter = ({ handleClick , state }) => {
   return (
-    <div className="flex justify-between items-center gap-10">
-      <div className="flex gap-20 justify-between">
+    <div className="flex justify-between">
+      <div className="flex justify-between w-full gap-1.5">
         <button disabled={state?.loading} onClick={() => handleClick("remove")}>
           <MinusSquare className={`${state?.loading ? "opacity-50" : null}`} />
         </button>
